Changes in details
In order to delete data by id
introduce Balanced tree with duplicate keys.
Feature achieved: insert / fetch
To-do: remove by id

+++ database/database.cpp
@@ -32,41 +32,8 @@ int conversion(datatype &data, unsigned int id, string datastr, string remark) {
 
 database::database(string databasePath)
 {
+	listTable(databasePath);
 	int strsize = databasePath.size();
-	//Time from http://blog.csdn.net/wangqing_12345/article/details/52092728
-	int stampTime = 0;
-	time_t tick = (time_t)stampTime;
-	struct tm tm;
-	char timestr[100];
-	Times standard;
-	//Traverse the folder from http://www.cnblogs.com/fnlingnzb-learner/p/6424563.html
-	string toSearch = databasePath.append("*.index"); databasePath.erase(strsize);
-	//const char *to_search = "C:\\database\\*.index";        //欲查找的文件，支持通配符
-	const char *to_search = toSearch.c_str();        //欲查找的文件，支持通配符
-
-	long handle;                                    //用于查找的句柄
-	struct _finddata_t fileinfo;                    //文件信息的结构体
-	handle = _findfirst(to_search, &fileinfo);      //第一次查找
-	if (-1 == handle) {
-
-	}
-	else {
-		printf_s("%s   ", fileinfo.name);				//打印出找到的文件的文件名
-		tick = fileinfo.time_access;
-		tm = *localtime(&tick);
-		strftime(timestr, sizeof(timestr), "%Y-%m-%d %H:%M:%S", &tm);
-		printf("Last modified on: %s\n", timestr);
-		while (!_findnext(handle, &fileinfo));			//循环查找其他符合的文件，直到找不到其他的为止
-		{
-			printf_s("%s   ", fileinfo.name);
-			tick = fileinfo.time_access;
-			tm = *localtime(&tick);
-			strftime(timestr, sizeof(timestr), "%Y-%m-%d %H:%M:%S", &tm);
-			printf("Last modified on: %s\n", timestr);
-		}
-		_findclose(handle);                              //别忘了关闭句柄
-	}
-
 	char table[40];
 	printf_s("Input the table name: "); scanf_s("%s", &table, 40);
 	string indexFileName = databasePath.append(table); indexFileName.append(".index"); databasePath.erase(strsize);
@@ -74,8 +41,9 @@ database::database(string databasePath)
 	string availableSpaceFileName = databasePath.append(table); availableSpaceFileName.append(".space");
 
 	bpt = new BPlusTree();
+	bpt_id = new BPlusTreePlus();
 	//db = new databaseIO(indexFileName, valueFileName, availableSpaceFileName, dataBPT, dataBPT_id);
-	db = new databaseIO(indexFileName, valueFileName, availableSpaceFileName, bpt, cache_insert, cache_remove);
+	db = new databaseIO(indexFileName, valueFileName, availableSpaceFileName, bpt, bpt_id);
 	printf_s("table %s selected successfully.\n", table);
 
 }
@@ -142,24 +110,13 @@ void database::insert(datatype &data, vector<dataBPTtype> &dataBPT, vector<dataB
 
 void database::insert(datatype &data)
 {//All the datas can be duplicated.
-	//cache_insert.push_back(data);
-	//if (cache_insert.size() > CACHESIZE) {
-	//	insert_update();
-	//}
 	int key = bpt->maxKey();
+	indexBPTtype indexBPTtypeTmp;
+	indexBPTtypeTmp.id = data.id;
+	indexBPTtypeTmp.key = key;
+	indexBPTtypeTmp.next = NULL;
+	bpt_id->insert(data.id, indexBPTtypeTmp);
 	bpt->insert(key, db->insert(key, data));
-}
-
-void database::insert_update()
-{
-	int key, i, j = cache_insert.size();
-	datatype datatypeTmp;
-	for (i = 0; i < j; i++) {
-		key = bpt->maxKey();
-		datatypeTmp = cache_insert.at(i);
-		bpt->insert(key, db->insert(key, datatypeTmp));
-	}
-	cache_insert.clear();
 }
 
 void database::modify(datatype &data, vector<dataBPTtype> &dataBPT, vector<dataBPTtype_id> &dataBPT_id) {
@@ -187,7 +144,7 @@ void database::remove(int id, vector<dataBPTtype> &dataBPT, vector<dataBPTtype_i
 void database::remove(unsigned int key)
 {
 	dataBPTtype dataBPTTmp;
-	if (bpt->search0(key, dataBPTTmp)) {
+	if (bpt->search(key) && bpt->search0(key, dataBPTTmp)) {
 		bpt->remove(key);
 		db->remove(dataBPTTmp);
 		printf_s("Successfully removed key=%d from database.\n", key);
@@ -197,9 +154,42 @@ void database::remove(unsigned int key)
 	}
 }
 
-void database::remove_update()
+void database::listTable(string databasePath)
 {
+	int strsize = databasePath.size();
+	//Time from http://blog.csdn.net/wangqing_12345/article/details/52092728
+	int stampTime = 0;
+	time_t tick = (time_t)stampTime;
+	struct tm tm;
+	char timestr[100];
+	Times standard;
+	//Traverse the folder from http://www.cnblogs.com/fnlingnzb-learner/p/6424563.html
+	string toSearch = databasePath.append("*.index"); databasePath.erase(strsize);
+	//const char *to_search = "C:\\database\\*.index";        //欲查找的文件，支持通配符
+	const char *to_search = toSearch.c_str();        //欲查找的文件，支持通配符
 
+	long handle;                                    //用于查找的句柄
+	struct _finddata_t fileinfo;                    //文件信息的结构体
+	handle = _findfirst(to_search, &fileinfo);      //第一次查找
+	if (-1 == handle) {
+
+	}
+	else {
+		printf_s("%s   ", fileinfo.name);				//打印出找到的文件的文件名
+		tick = fileinfo.time_access;
+		tm = *localtime(&tick);
+		strftime(timestr, sizeof(timestr), "%Y-%m-%d %H:%M:%S", &tm);
+		printf("Last modified on: %s\n", timestr);
+		while (!_findnext(handle, &fileinfo));			//循环查找其他符合的文件，直到找不到其他的为止
+		{
+			printf_s("%s   ", fileinfo.name);
+			tick = fileinfo.time_access;
+			tm = *localtime(&tick);
+			strftime(timestr, sizeof(timestr), "%Y-%m-%d %H:%M:%S", &tm);
+			printf("Last modified on: %s\n", timestr);
+		}
+		_findclose(handle);                              //别忘了关闭句柄
+	}
 }
 
 void database::get(int id, vector<dataBPTtype> &dataBPT, vector<dataBPTtype_id> &dataBPT_id) {
@@ -217,51 +207,33 @@ void database::get(int id, vector<dataBPTtype> &dataBPT, vector<dataBPTtype_id> 
 	}
 }
 
+void database::get(int id)
+{
+	vector<indexBPTtype> datas;
+	bpt_id->search0(id, datas);
+	int key;
+	//Method of iterator from http://blog.csdn.net/dgyanyong/article/details/21268469
+	vector<indexBPTtype>::iterator itor;
+	for (itor = datas.begin(); itor != datas.end(); itor++) {
+		printf_s("key     =%3u\n", itor->key);
+		printf_s("id      =%3u\n", itor->id);
+	}
+}
+
 void database::reopen(string databasePath) {
 	int strsize = databasePath.size();
-	//Time from http://blog.csdn.net/wangqing_12345/article/details/52092728
-	int stampTime = 0;
-	time_t tick = (time_t)stampTime;
-	struct tm tm;
-	char s[100];
-	Times standard;
-	//Traverse the folder from http://www.cnblogs.com/fnlingnzb-learner/p/6424563.html
-	string toSearch = databasePath.append("*.index"); databasePath.erase(strsize);
-	//const char *to_search = "C:\\database\\*.index";        //欲查找的文件，支持通配符
-	const char *to_search = toSearch.c_str();        //欲查找的文件，支持通配符
-
-	long handle;                                    //用于查找的句柄
-	struct _finddata_t fileinfo;                    //文件信息的结构体
-	handle = _findfirst(to_search, &fileinfo);      //第一次查找
-	if (-1 == handle) return;
-	printf_s("%s   ", fileinfo.name);				//打印出找到的文件的文件名
-	tick = fileinfo.time_access;
-	tm = *localtime(&tick);
-	strftime(s, sizeof(s), "%Y-%m-%d %H:%M:%S", &tm);
-	printf("Last modified on: %s\n", s);
-	while (!_findnext(handle, &fileinfo));			//循环查找其他符合的文件，直到找不到其他的为止
-	{
-		printf_s("%s   ", fileinfo.name);
-		tick = fileinfo.time_access;
-		tm = *localtime(&tick);
-		strftime(s, sizeof(s), "%Y-%m-%d %H:%M:%S", &tm);
-		printf("Last modified on: %s\n", s);
-	}
-	_findclose(handle);                              //别忘了关闭句柄
-
+	listTable(databasePath);
 	char table[40];
 	printf_s("Input the table name: "); scanf_s("%s", &table, 40);
 	string indexFileName = databasePath.append(table); indexFileName.append(".index"); databasePath.erase(strsize);
 	string valueFileName = databasePath.append(table); valueFileName.append(".value"); databasePath.erase(strsize);
 	string availableSpaceFileName = databasePath.append(table); availableSpaceFileName.append(".space");
-	db->reopen(indexFileName, valueFileName, availableSpaceFileName, dataBPT, dataBPT_id);
+	//db->reopen(indexFileName, valueFileName, availableSpaceFileName, dataBPT, dataBPT_id);
+	delete bpt;
+	delete bpt_id;
+	delete db;
+	bpt = new BPlusTree();
+	bpt_id = new BPlusTreePlus();
+	db = new databaseIO(indexFileName, valueFileName, availableSpaceFileName, bpt, bpt_id);
 	printf_s("table %s selected successfully.\n", table);
 }
-
-void database::flush()
-{
-	insert_update();
-	db->flush();
-}
-
-
